chore: refactored routes.py to use helper functions from helpers.py
